refactor: Refactors constants and stuff
