+++ components/home/features-section.tsx
@@ -6,7 +6,7 @@ import {
   FeatureCardIcon,
   FeatureCardTitle,
 } from '@/components/feature-card';
-import { featureItems } from '@/lib/constants';
+import { FEATURE_ITEMS } from '@/lib/constants';
 
 export default function FeaturesSection() {
   return (
@@ -14,7 +14,7 @@ export default function FeaturesSection() {
       <div className="py-30 w-300 max-w-full flex flex-col items-center gap-16">
         <h1 className="font-bold text-4xl text-white">Elevate your productivity to new heights</h1>
         <div className="grid grid-cols-3 gap-8 w-full">
-          {featureItems.map(({ title, content, Icon }, idx) => (
+          {FEATURE_ITEMS.map(({ title, content, Icon }, idx) => (
             <FeatureCard key={idx}>
               <FeatureCardCover>
                 <FeatureCardIcon>
